Persist auth tokens in cookies, and check the state of them when authenticating: * If the token is found and will not expire soon, immediately return the token. * If the token is found and may expire soon, authenticate via an iframe instead of showing the auto-closing popup. **EDIT** This has been removed from this CL; for now, the popup will always be used. * If the token is not found, assume access has not been granted and display the popup, which may auto-close.
The token is stored in cookies, as a mapping from
clientId + scopes -> token + expires

Since so much more is done in the Auth class, actually write tests for
it. They're JUnit tests, but I could write GWT tests for the
window-opening part too. Maybe later.

Update the Explorer to use the new Auth API -- this is pretty
straightforward.

Update the GWT client support code to use the new Auth API. Note that
this somewhat breaks the API of the GWT client, since it no longer
requires a CodeProvider implementation. Another option would be to
change WindowCodeProvider to call Auth.login() but that would require a
change to the CodeProvider interface to pass in the request parts
separately. I'm open to this but it seemed like a bigger change.

R=robertvawter
DELTA=1058  (821 added, 139 deleted, 98 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1499