Add skeleton of support for requesting private APIs from Discovery.
This will be available by adding the query parameter "tt" (e.g., "?tt"
in the URL), and will show a lock icon next to the logo, which, when
clicked, will show and OAuth2 prompt then reload the service selector
with the results of an authenticated request to Discovery.

As a short-term solution we ask Discovery if it knows its own auth
scope. If it does, and the "tt" parameter is given, we show the icon.
Otherwise, we know that Auth'd Discovery isn't there and we keep the
lock icon hidden.

R=jschorr
DELTA=148  (138 added, 6 deleted, 4 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2493